+++ src/main/java/me/tye/cogworks/CogWorks.java
@@ -303,8 +303,9 @@ private void automaticDependencyResolution() {
               ZipFile zip = new ZipFile(dependecyFile);
               for (Enumeration<? extends ZipEntry> e = zip.entries(); e.hasMoreElements(); ) {
                 ZipEntry entry = e.nextElement();
-                if (!entry.getName().equals("plugin.yml"))
+                if (!(entry.getName().equals("plugin.yml") || entry.getName().equals("paper-plugin.yml"))) {
                   continue;
+                }
 
                 StringBuilder out = new StringBuilder();
                 BufferedReader reader = new BufferedReader(new InputStreamReader(zip.getInputStream(entry)));